Pulled the 168kb moment.js out of textFormat.js (which is widely imported) so that we can make a separate smaller chunk which includes moment.js, this will allow us to only pull in moment when actually needed, and especially on the first page load.
