+++ src/js/components/Activity/ActivityTidbitComments.jsx
@@ -11,7 +11,7 @@ import ChildCommentList from './ChildCommentList';
 import ReactionStore from '../../stores/ReactionStore';
 import ReactionActions from '../../actions/ReactionActions';
 import { renderLog } from '../../utils/logging';
-import { timeFromDate } from '../../utils/textFormat';
+import { timeFromDate } from '../../utils/dateFormat';
 import VoterStore from '../../stores/VoterStore';
