+++ src/js/components/Share/SharedItemModal.jsx
@@ -24,7 +24,8 @@ import ShareActions from '../../actions/ShareActions';
 import SharedItemIntroduction from './SharedItemIntroduction';
 import ShareStore from '../../stores/ShareStore';
 import StepsChips from '../Widgets/StepsChips';
-import { convertToInteger, formatDateToMonthDayYear, startsWith } from '../../utils/textFormat';
+import { formatDateToMonthDayYear } from '../../utils/dateFormat';
+import { convertToInteger, startsWith } from '../../utils/textFormat';
 import VoterConstants from '../../constants/VoterConstants';
 import VoterStore from '../../stores/VoterStore';
 import VoterActions from '../../actions/VoterActions';
